Re-organised the mentor core tools. Not specific to #70, but a natural step.
