+++ src/ipbb/tools/mentor/__init__.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import
+
+from .sim_console import *
+from .sim_batch import *
